+++ Push-Swap/libps/ft_give_index.c
@@ -6,7 +6,7 @@
 /*   By: lchamard <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/13 15:18:12 by lchamard          #+#    #+#             */
-/*   Updated: 2026/01/13 15:18:23 by lchamard         ###   ########.fr       */
+/*   Updated: 2026/01/26 19:13:01 by lchamard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	give_index(t_list **stacka)
 	{
 		while (tmp && tmp->index > -1)
 			tmp = tmp->next;
-		while (tmp->next && !ft_lstisinf_ind(stacka, (tmp->value)))
+		while (tmp->next && !ft_lstisinf_ind(stacka, (long)(tmp->value)))
 		{
 			tmp = tmp->next;
 			while (tmp && tmp->index > -1)
